backend/requirements.txt
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
asyncpg==0.29.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.6
flower==2.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.34.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
kombu==5.5.4
langdetect==1.0.9
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
openai==1.106.1
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
postgrest==1.1.1
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
redis==6.4.0
requests==2.32.5
rsa==4.9.1
sentry-sdk==1.45.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.41.3
storage3==0.12.0
StrEnum==0.4.15
structlog==23.2.0
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
